Working finally :)

+++ CAR_SM/car_sm.c
@@ -6,7 +6,7 @@
  */ 
 
 #include "car_sm.h"
-extern uint32_t Distance;
+extern uint16_t Distance;
 ERROR_STATUS Car_SM_Init(void)
 {
 
@@ -20,14 +20,27 @@ ERROR_STATUS Car_SM_Update(void)
 {
 Us_Trigger();
 Us_GetDistance(&Distance);
+// if(Distance <= 40 && Distance >= 25 )
+// {
+// Steering_SteerCar(CAR_RIGHT,20);	
+// }
+// else if(Distance < 25)
+// {
+// Steering_SteerCar(CAR_BACKWARD,20);
+// 
+// }else if(Distance > 40) 
+// {
+// 	Steering_SteerCar(CAR_FORWARD,20);
+// }	
+
 if(Distance <= 30)
 {
 	
-Steering_SteerCar(CAR_RIGHT,30);	
+	Steering_SteerCar(CAR_RIGHT,30);
 }else
 {
 	Steering_SteerCar(CAR_FORWARD,30);
-}	
+}
 }
